Refactor: tests: create separate `robots.txt`-related test module
